% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b-status.R
\name{b_status}
\alias{b_status}
\title{Did not bat function.}
\usage{
b_status(status)
}
\arguments{
\item{status}{A number, DNB, or TDNB.}
}
\description{
Convert bowling/batting batting category to a character variable.
}
