% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-status.R
\name{create_status}
\alias{create_status}
\title{Did not bat function.}
\usage{
create_status(status)
}
\arguments{
\item{status}{A number, DNB, or TDNB.}
}
\description{
Convert bowling/batting batting category to a character variable.
}
