% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricinfo-package.R
\docType{package}
\name{cricinfo}
\alias{cricinfo}
\alias{cricinfo-package}
\title{Functions for Downloading Cricket Data from ESPNCricinfo}
\description{
Data on all international cricket matches is provided by ESPNCricinfo.
This package provides some scraper functions to download the data 
into tibbles ready for analysis.
}
\details{
\tabular{ll}{ Package: \tab cricinfo\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\keyword{package}
