% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricketdata.R
\docType{package}
\name{cricketdata-package}
\alias{cricketdata}
\alias{cricketdata-package}
\title{cricketdata: International Cricket Data}
\description{
Data on international and other major cricket matches from ESPNCricinfo \url{https://www.espncricinfo.com} and Cricsheet \url{https://cricsheet.org}. This package provides some functions to download the data into tibbles ready for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/cricketdata/}
  \item \url{https://github.com/robjhyndman/cricketdata}
  \item Report bugs at \url{https://github.com/robjhyndman/cricketdata/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu}

Authors:
\itemize{
  \item Charles Gray \email{C.Gray@latrobe.edu.au}
  \item Sayani Gupta \email{Sayani.Gupta@monash.edu}
  \item Timothy Hyndman \email{Timothy.Hyndman@gmail.com}
  \item Hassan Rafique \email{dazzalytics@protonmail.com}
  \item Jacquie Tran \email{jac@jacquietran.com}
}

Other contributors:
\itemize{
  \item Puwasala Gamakumara \email{Puwasala.Gamakumara@monash.edu} [contributor]
  \item Alex Whan \email{alexwhan@gmail.com} [contributor]
}

}
\keyword{internal}
