% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricketdata.R
\docType{package}
\name{cricketdata}
\alias{cricketdata}
\alias{cricketdata-package}
\title{Cricket Data from International Matches}
\description{
Data on all international cricket matches is provided by ESPNCricinfo.
This package provides some scraper functions to download the data 
into tibbles ready for analysis. 
Please respect the 
\href{http://www.espncricinfo.com/ci/content/site/company/terms_use.html}{ESPNCricinfo terms of use}.
Some innings-level data sourced from Howzstat (\href{http://www.cricket-stats.net/}) 
is also included in the package.
}
\details{
\tabular{ll}{ Package: \tab cricketdata\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\keyword{package}
