% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cricsheet_codes}
\alias{cricsheet_codes}
\title{Codes used for competitions on Cricsheet}
\format{
A data frame with 44 rows and 2 variables.
}
\source{
\url{https://cricsheet.org/downloads/#experimental}
}
\usage{
cricsheet_codes
}
\description{
A dataset containing the names and codes used by
\href{https://cricsheet.org}{cricsheet}, as at 24 March 2025.
}
\keyword{datasets}
