% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_batting_career.R
\name{fetch_batting_career}
\alias{fetch_batting_career}
\title{Fetch and clean batting career dataset.}
\usage{
fetch_batting_career(matchtype = c("test", "odi", "t20"), sex = c("men",
  "women"))
}
\arguments{
\item{matchtype}{Character indicating test, odi, or t20.}

\item{sex}{Character indicating men or women.}
}
\description{
Retrieve career batting data for each person on ESPNCricinfo
}
\examples{
wt20_batting_career <- fetch_batting_career("t20", "women")

}
