% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_batting_innings.R
\name{fetch_batting_innings}
\alias{fetch_batting_innings}
\title{Fetch and clean batting innings dataset.}
\usage{
fetch_batting_innings(matchtype = c("test", "odi", "t20"), sex = c("men",
  "women"))
}
\arguments{
\item{matchtype}{Character indicating test, odi, or t20.}

\item{sex}{Character indicating men or women.}
}
\description{
Retrieve batting data for every innings from ESPNCricinfo
}
\examples{
wt20_batting_innings <- fetch_batting_innings("t20", "women")

}
