% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_bowling_innings.R
\name{fetch_bowling_innings}
\alias{fetch_bowling_innings}
\title{Get and clean bowling innings dataset.}
\usage{
fetch_bowling_innings(matchtype = c("test", "odi", "t20"), sex = c("men",
  "women"))
}
\arguments{
\item{matchtype}{Character indicating test, odi, or t20.}

\item{sex}{Character indicating men or women.}
}
\description{
Retrieve bowling data for innings from the
}
\examples{
bowling_data <- fetch_bowling_innings("t20", "women")

}
