% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cricsheet_bbb.R
\name{fetch_cricsheet}
\alias{fetch_cricsheet}
\title{Fetch Cricsheet ball-by-ball, match and player data}
\usage{
fetch_cricsheet(
  competition = c("apl", "bbl", "bpl", "cpl", "county", "edwards_cup",
    "heyhoe_flint_trophy", "ipl", "multi_day_matches", "msl", "odi_men", "odi_women",
    "odm_men", "odm_women", "psl", "sheffield_shield", "super_smash_men",
    "super_smash_women", "tests_men", "tests_women", "the_hundred_men",
    "the_hundred_women", "t20_blast", "t20_int_men", "t20_int_women",
    "t20_non_official_men", "t20_non_official_women", "wbbl", "wipl"),
  type = c("bbb", "match", "player")
)
}
\arguments{
\item{competition}{Character string giving name of competition.}

\item{type}{Character string giving type of data: ball-by-ball, match info or player info.}
}
\description{
Download data from Cricsheet \url{cricsheet.org}.
}
\examples{
\dontrun{
wbbl_bbb <- fetch_cricsheet_bbb("wbbl")
}
}
\author{
Jacquie Tran and Rob J Hyndman
}
