% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_player_data.R
\name{fetch_player_data}
\alias{fetch_player_data}
\title{Fetch Player Data}
\usage{
fetch_player_data(playerid, matchtype = c("Test", "ODI", "T20"))
}
\arguments{
\item{playerid}{The player ID as given in the Cricinfo profile.}

\item{matchtype}{Which type of cricket matches do you want? Tests, ODIs or T20s?}
}
\value{
A tibble containing data on the selected player, with one row for every innings
of every match in which they have played.
}
\description{
Fetch individual player data from all matches played. The function will scrape
the data from ESPNCricinfo and return a tibble with one line per innings for all
games a player has played. To identify a player, use their Cricinfo player ID.
The simplest way to find this is to look up their Cricinfo Profile page. The number
at the end of the URL is the ID. For example, Meg Lanning's profile page is
http://www.espncricinfo.com/australia/content/player/329336.html,
so her ID is 329336.
}
\examples{
ElyssePerry <- fetch_player_data(275487, "T20")
MegLanning <- fetch_player_data(329336, "ODI")
SteveSmith <- fetch_player_data(267192, "Test")
library(ggplot2)
ggplot(MegLanning) + geom_point(aes(x=Date, y=Score, col=NotOut)) +
  ggtitle("Meg Lanning ODI Scores")

}
\author{
Rob J Hyndman
}
