% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_player_id.R
\name{find_player_id}
\alias{find_player_id}
\title{Find a player id from cricinfo.com}
\usage{
find_player_id(searchstring)
}
\arguments{
\item{searchstring}{Part of a player name to search for.}
}
\value{
A table of matching players, their ids, and teams they played for.
}
\description{
Find a player id from cricinfo.com
}
\examples{
perry <- find_player_id("Perry")
ElyssePerry <- fetch_player(perry[2,1])
}
