% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participation_status.R
\name{participation_status}
\alias{participation_status}
\title{Did not bat function.}
\usage{
participation_status(status)
}
\arguments{
\item{status}{A number, DNB, or TDNB.}
}
\description{
Convert bowling/batting batting category to a character variable.
}
