% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_meta}
\alias{player_meta}
\title{Meta data on all players listed at ESPNCricinfo}
\format{
A data frame with 11342 rows and 8 variables.
}
\source{
\url{https://www.espncricinfo.com}
}
\usage{
player_meta
}
\description{
A dataset containing the names and other attributes of players
with profiles on ESPNCricinfo, as at 3 August 2022.
}
\keyword{datasets}
