% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cricketdata.R
\docType{package}
\name{cricketdata}
\alias{cricketdata}
\title{Cricket Data from international and other major cricket matches.}
\description{
Data is provided by ESPNCricinfo and Cricsheet.
This package provides some functions to download the data 
into tibbles ready for analysis. 
Please respect the terms of use for each website: 
\href{https://www.espncricinfo.com/ci/content/site/company/terms_use.html}{ESPNCricinfo},
\href{https://cricsheet.org/register/}{Cricsheet}.
}
\details{
\tabular{ll}{Package: \tab cricketdata\cr Type: \tab Package\cr License: \tab
GPL3\cr LazyLoad: \tab yes\cr }
}
\keyword{package}
