% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cricket_data.R
\name{fetch_cricket_data}
\alias{fetch_cricket_data}
\title{Get cricket data.}
\usage{
fetch_cricket_data(matchtype = c("test", "odi", "t20"), sex = c("men",
  "women"), activity = c("batting", "bowling", "fielding"),
  view = c("innings", "career"), team = NULL)
}
\arguments{
\item{matchtype}{Character indicating test, odi, or t20.}

\item{sex}{Character indicating men or women.}

\item{activity}{Character indicating batting, bowling, or fielding.}

\item{view}{Character indicating innings or career.}
}
\description{
Retrieve cricket data from stats cricinfo based on various categories.
}
\examples{
wt20_bowling <- fetch_cricket_data("t20", "women", "bowling", "career")
}
