% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_player_id.R
\name{find_player_id}
\alias{find_player_id}
\title{Find a player id from cricinfo.com}
\usage{
find_player_id(searchstring)
}
\arguments{
\item{searchstring}{Part of a player name(s) to search for. Can be a character vector.}
}
\value{
A table of matching players, their ids, and teams they played for.
}
\description{
Find a player id from cricinfo.com
}
\examples{
(perry <- find_player_id("Perry"))
EllysePerry <- fetch_player_data(perry[2,"ID"], "test")
}
\author{
Rob J Hyndman
}
