% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_meta}
\alias{player_meta}
\title{Meta data on players listed at ESPNCricinfo}
\format{
A data frame with 16101 rows and 11 variables.
}
\source{
\url{https://www.espncricinfo.com}
}
\usage{
player_meta
}
\description{
A dataset containing the names and other attributes of players who appear
on both \href{https://cricsheet.org}{cricsheet} and
\href{https://www.espncricinfo.com}{ESPNCricinfo}, as at 24 March 2025.
}
\keyword{datasets}
